split http implementations into curl- and wininet-based
