+++ http.h
@@ -30,23 +30,15 @@ struct HttpHeader {
 struct HttpResponse {
   HttpStatus status;
   std::vector<HttpHeader> headers;
-  std::vector<uint8_t> body;
+  std::vector<std::uint8_t> body;
 };
 
-struct CurlDeleter {
-  void operator()(void*) const;
-};
-
-class HttpGet {
-  std::vector<HttpHeader> headers;
-  std::string resource;
-  std::unique_ptr<void, CurlDeleter> curl;
-
-public:
-  HttpGet();
-  void add_header(const std::string& name, const std::string& value);
-  void set_resource(const std::string& res);
-  bool send(const std::string& hostname, HttpResponse& out) const;
-};
+#ifdef _WIN32
+#include "http_wininet.h"
+using HttpGet = HttpGetWinInet;
+#else
+#include "http_curl.h"
+using HttpGet = HttpGetCurl;
+#endif
 
 #endif
